konflux: don't pass prefetch artifact wastefully
The sast-shell-check and sast-unicode-check tasks do not use the
prefetched content in any way, even though they do, for some reason,
accept a CACHI2_ARTIFACT parameter and *will* fetch the artifact if
passed in.

In our case, the prefetched artifact is huge, so passing it to tasks
that don't use it is a massive waste of time.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>