devtool: determine repo root differently
The previous approach could have confusing results in some cases. The
__file__ variable always points to the file in the virtualenv you
currently have active. When you, for example, create a git worktree in a
separate directly, devtool would still be running against the original
directory in which you created the venv.

Get the repo root with 'git rev-parse --show-toplevel' instead.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>